@choleraehyq 2020-09-20
Benchmark output in the commit message is using benchstat now.